tema6

+++ app/src/main/java/ro/ase/moneysaver/Buget.java
@@ -4,17 +4,17 @@ import java.io.Serializable;
 
 public class Buget implements Serializable
 {
-    private double suma;
+    private float suma;
 
-    public Buget(double suma) {
+    public Buget(float suma) {
         this.suma = suma;
     }
 
-    public double getSuma() {
+    public float getSuma() {
         return suma;
     }
 
-    public void setSuma(double suma) {
+    public void setSuma(float suma) {
         this.suma = suma;
     }
